Added endpoint for extracting comments from zipfile
