+++ skema/skema-rs/comment_extraction/src/lib.rs
@@ -2,4 +2,5 @@
 
 pub mod comments;
 pub mod conventions;
+pub mod extraction;
 pub mod languages;
