+++ skema/skema-rs/skema/Cargo.toml
@@ -22,3 +22,7 @@ petgraph = "0.6.2"
 clap = { version = "4.0.26", features = ["derive"] }
 utoipa-swagger-ui = { version = "3.0.2", features = ["actix-web"] }
 schemars = { version = "0.8.12" }
+tempdir = "0.3.7"
+log = "0.4.19"
+pretty_env_logger = "0.5.0"
+zip = "0.6.6"
