+++ skema/skema-rs/skema/src/bin/skema_service.rs
@@ -34,10 +34,13 @@ pub async fn ping() -> HttpResponse {
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
+    pretty_env_logger::init();
+
     #[derive(OpenApi)]
     #[openapi(
         paths(
             comment_extraction::get_comments,
+            comment_extraction::get_comments_from_zipfile,
             skema::services::mathml::get_ast_graph,
             skema::services::mathml::get_math_exp_graph,
             skema::services::mathml::get_acset,
@@ -121,6 +124,7 @@ async fn main() -> std::io::Result<()> {
             }))
             .configure(gromet::configure())
             .service(comment_extraction::get_comments)
+            .service(comment_extraction::get_comments_from_zipfile)
             .service(skema::services::mathml::get_ast_graph)
             .service(skema::services::mathml::get_math_exp_graph)
             .service(skema::services::mathml::get_acset)
