+++ skema/skema-rs/skema/src/services/comment_extraction.rs
@@ -1,11 +1,14 @@
 //! Comment extraction services
 
-use actix_web::{get, web, HttpResponse};
-use comment_extraction::comments::Comments;
+use actix_web::{get, post, web, HttpResponse};
+use comment_extraction::{comments::Comments, extraction::extract_comments_from_directory};
 use comment_extraction::languages::python::get_comments_from_string as get_python_comments;
 use serde::{Deserialize, Serialize};
 use utoipa;
 use utoipa::ToSchema;
+use std::{fs, io::Write};
+use log::debug;
+use tempdir::TempDir;
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub enum Language {
@@ -76,4 +79,33 @@ impl CommentExtractionRequest {
 #[get("/get_comments")]
 pub async fn get_comments(payload: web::Json<CommentExtractionRequest>) -> HttpResponse {
     HttpResponse::Ok().json(web::Json(get_python_comments(&payload.code)))
+}
+
+
+/// Get comments for a zipfile containing a directory of code
+#[utoipa::path(
+    responses(
+        (status = 200, description = "Get comments from zipfile")
+    )
+)]
+#[post("/get-comments-from-zipfile")]
+pub async fn get_comments_from_zipfile(payload: web::Bytes) -> HttpResponse {
+    debug!("Zip file received. Size: {} bytes", payload.len());
+    let tmp_dir = TempDir::new("comment_extraction_input").unwrap();
+    let file_path = tmp_dir.path().join("directory.zip");
+    //let file_path = "test.zip";
+    let mut f = fs::File::create(file_path.clone()).unwrap();
+    f.write_all(&payload).unwrap();
+    drop(f);
+
+    // Read zipfile
+    let file = fs::File::open(file_path).unwrap();
+    let mut archive = zip::ZipArchive::new(file).unwrap();
+    archive.extract(&tmp_dir).unwrap();
+    let comments = extract_comments_from_directory(&tmp_dir, &None);
+
+    tmp_dir.close().unwrap();
+
+    HttpResponse::Ok().json(web::Json(comments))
+    //HttpResponse::Ok().finish()
 }
